:truck: move(View, Controller) movendo classes para pastas para com suas respectivas funções
