+++ src/main/java/com/agibank/corehub/controller/home/LoginController.java
@@ -57,6 +57,10 @@ public class LoginController {
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/com/agibank/corehub/views/home.fxml"));
         Parent root = loader.load();
 
+        HomeController homeController = loader.getController();
+        homeController.labelNomeUsuario.setText(usuario.getNome());
+        homeController.setIdUsuario(usuario.getId_Usuario());
+
         Stage stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
         Scene scene = new Scene(root, 412, 915);
         stage.setScene(scene);
